broke rotation by trying to pass it from rectangle to abstract shape - check previous commit to see whats going wrong

+++ src/main/java/org/herebdragons/graphics/objects/Movable.java
@@ -10,7 +10,7 @@ public interface Movable {
 
     void move(int x, int y);
 
-    void move(float x, float y);
+    void move(double x, double y);
 
 
 }
